+++ buyables/vehicles/spe_us_late.hpp
@@ -1,0 +1,22 @@
+class Vehicles: vehicles_base {
+
+    class Box_NATO_Equip_F: Box_NATO_Equip_F {};
+
+	class SPE_US_M3_Halftrack_Unarmed_Open {
+        condition = "true";
+        price = 1000;
+        stock = 10;
+    };
+
+    class SPE_US_M3_Halftrack {
+        condition = "true";
+        price = 3500;
+        stock = 10;
+    };
+
+    class SPE_M4A1_76_DLV {
+        condition = "true";
+        price = 5500;
+        stock = 10;
+    };
+};
