--- buyables/vehicles/spe_wehrmacht.hpp
@@ -1,22 +1,0 @@
-class Vehicles: vehicles_base {
-
-    class Box_NATO_Equip_F: Box_NATO_Equip_F {};
-
-	class SPE_OpelBlitz {
-        condition = "true";
-        price = 1000;
-        stock = 10;
-    };
-
-    class SPE_SdKfz250_1 {
-        condition = "true";
-        price = 4500;
-        stock = 10;
-    };
-
-    class SPE_PzKpfwIII_J {
-        condition = "true";
-        price = 5500;
-        stock = 10;
-    };
-};
